add new update Blue-Line-OE-4ATV

+++ Blue-Line-OE-4ATV/usr/lib/enigma2/python/Components/Renderer/BLWatches.py
@@ -1,6 +1,5 @@
-# Embedded file name: /usr/lib/enigma2/python/Components/Renderer/AXBlueWatches.py
 import math
-from Renderer import Renderer
+from Components.Renderer.Renderer import Renderer
 from skin import parseColor
 from enigma import eCanvas, eSize, gRGB, eRect
 from Components.VariableText import VariableText
@@ -39,7 +38,7 @@ class BLWatches(Renderer):
     def hand(self):
         width = self.instance.size().width()
         height = self.instance.size().height()
-        r = min(width, height) / 2
+        r = int(min(width, height) / 2)
         endX, endY = self.calculate(self.numval, r, r)
         self.draw_line(r, r, endX, endY)
